refactor
adjust Git sink to fit into to conform to project convetions
