+++ gix-diff/tests/diff/blob/mod.rs
@@ -1,3 +1,4 @@
+mod git_diff;
 pub(crate) mod pipeline;
 mod platform;
 mod unified_diff;
